0.3.0-alpha Code updates
Next batch of code updates for the upcoming pre-release version 0.3.0-alpha (Pre-Alpha 3.0). Updates include increasing height of the window from 500px to 550PX, making copyright year auto update to the current year using new date class, added an option to add a title to a game of name that tune, as well as still in progress/uncompleted code for the new Save feature.