0.3.0-alpha code update
Did more coding for pre-release version 0.3.0, work done includes some work on the methods in the Save class (including being able to open Save Dialog Box), added validation checks to the title textfield on the game setup page in order to prevent characters that are not allowed on the file path if title is used as part of a file. Also modified the validation check for the number of players textfield in GameSetupController class in order to prevent it from clashing/overriding the checks to the title field. 

More work to come soon!