0.3.0-alpha Update
Completed code for 0.3.0-alpha. This pre-release version includes 2 new features including the new 'Save' feature which will allow users to save game results to text file so they can  view it later. A minor new feature was also added that allows users to add an optional title for a game of Name That Tune (ex. 90's edition, 00's edition, etc), this title will be displayed on the results screen when the game is finished, the title will also be used to create a default file name when using the save feature. Other minor improvements and changes include increasing the height of the window from 500px to 550px, adding the tie breaker song to the 'Select Winner' tiebreaker mode screen and also moving the 'Selected Song' and 'Selected Player' labels that are in the 'Set Players Scores' screen to their own lines and making them bold so they stand out better. 

A complete and detailed explanation of each change (including code base changes) will be available in the full release notes for this version.