Source Code for 0.3.1-alpha
Completed source code for version 0.3.1-alpha. Changes in this version include an edit feature that lets the user edit songs and players during the setup process, a scoring history on the scoring screen along with a new undo feature for undoing previous scoring changes. Additionally instructions for song and player setup screens were redone with a new tabbed layout and simplified instructions. 

Code for several files was also refactored and cleaned up to improve readability and maintainability of the code base as development continues.